backtest/ai_trend/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.25.0
supabase>=1.0.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
ccxt>=4.0.0 